feat: ✨ lots of additions and code cleanup
feat: ✨ luarocks

add:
- luarocks PackageManager
- rocks subcommand to install packages

feat: ✨ add a customizable string expander

refactor: 🔨 cleanup a lot of code

feat: ✨ add some more lua bindings

feat: ✨ add a better event bus

feat: ✨ add a new jsonschema validator

feat: ✨ add OrchestratorRunMode enum

feat: ✨ initial plan implementation

add:
- plan command now runs end to end
- plan output is prettified using lipgloss
- controllers can now leave a reason when running plan

feat: ✨ add some basic manifest bindings utils for lua

refactor: 🔨 working on adding an emit function to lua for events

refactor: 🔨 add a lot of code cleanup

refactor: 🔨 make the validation reason a stack allocated char*

feat: ✨ add telemetry tracking for the orchestrator

refactor: 🔨 cleanup OrchestratorMetrics

refactor: 🔨 convert history to use Reason, break the history subcommand

refactor: 🔨 add a TODO

refactor: 🔨 add another TODO

build: 📦 add uuid dep to core lib

refactor: 🔨 cleanup the ResourceGraph code a bit

refactor: 🔨 more code cleanup

fix: 🐛 fix issue with freeing stack allocated bitset

refactor: 🔨 more code cleanup

refactor: 🔨 more code cleanup

refactor: 🔨 cleanup some of the controllers code

refactor: 🔨 cleanup cli code

refactor: 🔨 add more dummy package manager implementations

refactor: 🔨 fix building with the newly added dummy package managers

refactor: 🔨 break something

feat: ✨ add basic implementation for lua controllers

build: 📦 cleanup meson builds a abit

refactor(lua-runtime): 🔨 cleanup lua runtime code a bit now that its working

refactor(bindings/events): 🔨 cleanup bindings

refactor(bindings/template): 🔨 cleanup bindings

refactor(bindings/jsonnet): 🔨 cleanup bindings

refactor(bindings/log): 🔨 cleanup bindings

refactor(bindings/manifest): 🔨 cleanup bindings

refactor(bindings/packages): 🔨 cleanup bindings

refactor(bindings/resource): 🔨 cleanup bindings

refactor(bindings/runtime): 🔨 cleanup bindings

+++ libs/resource-graph/include/hypha/resource_graph.h
@@ -6,7 +6,7 @@
 #include "hypha/resource.h"
 #include "hypha/resource_selector.h"
 
-typedef int32_t ResourceGraphIndex;
+typedef size_t ResourceGraphIndex;
 static const ResourceGraphIndex kInvalidResourceIndex = -1;
 
 typedef struct _ResourceGraph ResourceGraph;
